re-organize, add panic
