+++ lib/tmp.test.sql.js
@@ -1,4 +1,3 @@
-
 (function(){
 	if(!this.Gun){ return }
 	function Test(o){
@@ -239,24 +238,28 @@ Gun.put.wire.from.parse = function(t){
 }
 }());
 */
-/*
+
 ;(function(){ // make as separate module!
+	function SQL(){}
+	SQL.select = function(sel){
+		this._.sql.select = sel;
+		return this;
+	}
+	SQL.from = function(from){
+		this._.sql.from = from;
+		//this.get(from).map();
+		return this;
+	}
+	SQL.where = function(where){
+		this._.sql.where = where;
+		return this;
+	}
 	Gun.chain.sql = function(sql){
 		var gun = this;//.chain();
 		sql = gun._.sql = sql || {};
-		gun.select = function(sel){
-			sql.select = sel;
-			return gun;
-		}
-		gun.from = function(from){
-			sql.from = from;
-			gun.get(from).map();
-			return gun;
-		}
-		gun.where = function(where){
-			sql.where = where;
-			return gun;
-		}
+		gun.select = SQL.select;
+		gun.from = SQL.from;
+		gun.where = SQL.where;
 		return gun;
 	}
 
@@ -267,4 +270,3 @@ Gun.put.wire.from.parse = function(t){
 		var node = at.node;
 	});
 }());
-*/
