+++ test/mocha.html
@@ -16,7 +16,7 @@
 		<script>mocha.setup({ui:'bdd',globals:[]});</script>
 		<script src="./expect.js"></script>
 		<script></script>
-		<script src="../../gun.js"></script>
+		<script src="../gun.js"></script>
 		<script src="./common.js"></script>
 		<script>mocha.run()</script>
 	</body>
