+++ test/ptsd/ptsd.html
@@ -31,7 +31,7 @@
 	<script src="https://code.jquery.com/jquery-1.12.4.min.js"></script>
 	<script src="benchmark.js"></script>
 	<script src="ptsd.js"></script>
-	<script src="../gun.js"></script>
+	<script src="../../gun.js"></script>
 	<script src="perf.js"></script>
 </body>
 </html>
